+++ framework-docs/modules/ROOT/nav.adoc
@@ -381,7 +381,7 @@
 **** xref:testing/mockmvc/htmlunit/mah.adoc[]
 **** xref:testing/mockmvc/htmlunit/webdriver.adoc[]
 **** xref:testing/mockmvc/htmlunit/geb.adoc[]
-*** xref:testing/mockmvc/playwright/playwright.adoc[]
+*** xref:testing/mockmvc/playwright.adoc[]
 *** xref:testing/mockmvc/vs-end-to-end-integration-tests.adoc[]
 *** xref:testing/mockmvc/resources.adoc[]
 ** xref:testing/spring-mvc-test-client.adoc[]
