+++ framework-docs/modules/ROOT/pages/testing/mockmvc/playwright.adoc
@@ -10,8 +10,10 @@ starting a server.
 == Setup
 
 First, make sure that you have included a test dependency on `com.microsoft.playwright:playwright`.
-To use it, create a `MockMvc` instance and register `MockMvcPlaywrightHandler` as a
-route https://playwright.dev/java/docs/network#handle-requests[handler] for requests.
+To use it, create a`MockMvcPlaywrightHandler` instance and register it as a
+route https://playwright.dev/java/docs/network#handle-requests[handler] for http requests.
+
+include-code::./MockMvcPlaywrightHandlerTests[tag=init,indent=0]
 
 include-code::./MockMvcPlaywrightHandlerTests[tag=setup,indent=0]
